mach-mxs: Added exception for TS-7680 to force enable enet clock early
This is required for setup of the VLAN mode. In order to configure the
switch IC to have VLAN ports, the clock needs to be up and active. Ideally
VLAN mode would be set before the interface is brought up which would mean
there is no active clock. This patch forces the clock on earlier in boot.