net: fec: Workaround to add hardware PHY reset to fec_restart()
Current FEC driver makes the assumption that its okay to turn the PHY
clock off and turn it back on and expect the PHY to behave.

e.g. 'ifconfig ethX down && ifconfig ethX up' While the interface is
down the clock is simply turned off. When it comes back up, the clock
is turned on and a software reset is issued. This is not enough and a
hardware reset is required. However FEC is not set up to issue
arbitrary PHY hardware resets and only does it at probe().

Modification is a workaround to allow fec_reset_phy() to be called
multiple times during the life of the driver. fec_restart() will now
call fec_reset_phy() at the end of its reconfiguration to allow the new
settings to take effect.