spi: oc: Clear automatic CS control before transfer
CTRL_ASS is automatic slave select. While useful in some situations,
this driver expects it to not be enabled and instead have manual CS
control. On some TS platforms, CTRL_ASS may be used in U-Boot, and
must be disabled in the peripheral before the driver can properly
manage the FPGA core.