ARM: ts_imx28_defconfig: ChipIdea and OTG now built-in
The ChipIdea Highspeed Dual Role Controller driver and the OTG
FSM drivers are required for USB operation either as host or
device. Previously set as modules, this prevented booting directly
to a USB mass storage device. These drivers are now set to be
built-iin to the kernel rather than as modules.