watchdog: ts7100_wdt: Added initial driver
The TS-7100 (and other devices) are no longer compatible with the
existing ts_wdt driver. This 2nd generation supports a more
traditional WDT implementation with separate timeout and feed regs.
The WDT is still contained inside the supervisory microcontroller
and is set up to be an I2C driver with WDT support.