workflows: Initial commit of build test
Trigger an attempt to build the kernel when a new tag is pushed.
The tag syntax is filtered on 'v*-ts' to prevent this job running
if upstream kernel tags accidentally ever get pushed. Upstream
kernel uses 'v*' without the trailing -ts

Triggers a build attempt as well on an opened PR (This may want to
be removed if there is a sudden flood of PRs since that would eat
our "actions" time up)

Note that this only builds with the tsimx6ul_defconfig for now