tty: serial: imx: Honor LOW_LATENCY ioctl
When low_latency is set, DMA is not enabled and latency is reduced

This is needed when using RS-485 as there is a bug that exists,
even upstream, where TXEN stays asserted and transmits repeated
NULL bytes.

Note that with low_latency set, the UARTs are not capable of going
above 115200 baud reliably and may lose data.