serial: core: Clear circ buffer on xmit.buf reuse
This fixes an issue where a 16550 that is opened, flooded with writes, then closed, will
still have data remaining in the circ buffer. When the uart starts up, it blocks on
tty_chars_in_buffer to drain. Since nothing at this point in startup is pulling
data from this circ_buffer, the uart sits indefinitely.