ARM: dts: TS: Disable DMA on RS-485 ports
This is a compromise for an issue where with DMA enabled, the
TXEN line can get asserted for long periods of time with either
NULL or the last byte being repeatedly transmitted during this
time. Disabling DMA limits the RS-485 ports to 115200 baud max
before the RX buffers may start to overflow during periods of
maximum data rates.